grafana-pdf-report-generator/requirements.txt
azure-communication-email
azure-identity